refactor: cleanup dependencies & redesign config
